Move serializing logic from initializer to existing OpenApi namespace
